+++ src/components/file/File.js
@@ -298,6 +298,7 @@ export default class FileComponent extends BaseComponent {
       },
       class: 'browse'
     }, this.text('browse'));
+    this.addFocusBlurEvents(this.browseLink);
 
     return this.browseLink;
   }